--- Sample.Core/Models/Cart.cs
@@ -1,9 +1,0 @@
-using System.Collections.ObjectModel;
-
-namespace Sample.Core.Models
-{
-    public class Cart
-    {
-        public ObservableCollection<Item> LineItems { get; set; } = new();
-    }
-}
