+++ Sample.Core/Models/TaxJar/Orders/LineItem.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace Sample.Core.Models.Orders
+namespace Sample.Core.Models.TaxJar.Orders
 {
     public class LineItem
     {
