+++ Sample.Core/Models/TaxJar/Rates/RatesResponse.cs
@@ -1,7 +1,7 @@
 using System.Net.Http;
 using Newtonsoft.Json;
 
-namespace Sample.Core.Models.Rates
+namespace Sample.Core.Models.TaxJar.Rates
 {
     public class RatesResponse : HttpResponseMessage
     {
