+++ Sample.Core/Models/TaxJar/Taxes/Jurisdictions.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace Sample.Core.Models.Taxes
+namespace Sample.Core.Models.TaxJar.Taxes
 {
     public class Jurisdictions
     {
