+++ Sample.Core/Resources/ViewLabels/StoreViewModel.json
@@ -1,0 +1,3 @@
+{
+  "Title": "Store"
+}
