--- Sample.Core/Resources/ViewLabels/WelcomeViewModel.json
@@ -1,4 +1,0 @@
-{
-  "Title": "Home",
-  "WelcomeButton": "Sample List View"
-}
