+++ Sample.Core/Views/StorePage.xaml
@@ -6,12 +6,13 @@
                       xmlns:viewModels="clr-namespace:Sample.Core.ViewModels"
                       xmlns:views="clr-namespace:MvvmCross.Forms.Views;assembly=MvvmCross.Forms"
                       xmlns:mvxViewModels="clr-namespace:MvvmCross.ViewModels;assembly=MvvmCross"
-                      x:Class="Sample.Core.Views.WelcomePage"
-                      x:DataType="viewModels:WelcomeViewModel"
-                      x:TypeArguments="viewModels:WelcomeViewModel">
+                      xmlns:taxCalcStore="clr-namespace:Sample.Core.Models.TaxCalcStore;assembly=Sample.Core"
+                      x:Class="Sample.Core.Views.StorePage"
+                      x:DataType="viewModels:StoreViewModel"
+                      x:TypeArguments="viewModels:StoreViewModel">
     <ContentPage.Resources>
         <DataTemplate x:Key="ItemsTemplate"
-                      x:DataType="models:Item">
+                      x:DataType="taxCalcStore:Item">
             <ContentView>
                 <controls:ShopFrame Item="{Binding .}"
                                     Margin="8"
@@ -24,9 +25,9 @@
         <StackLayout>
             <controls:NavBar Title="{Binding [Title]}"
                              SafeAreaInsetEffectEnabled="True"
-                             NavigationState="Close"
                              NavigationCommand="{Binding BackCommand}"
                              QuickActionState="Shop"
+                             QuickActionCommand="{Binding MyCartCommand}"
                              QuickActionDotLabel="{Binding Cart.LineItems.Count}"/>
             <ScrollView>
                 <CollectionView ItemsSource="{Binding Items}"
